remove starknet_ preffix from files (#588)
