+++ src/business_logic/transaction/objects/internal_declare.rs
@@ -12,9 +12,8 @@ use crate::{
         },
     },
     core::{
-        contract_address::starknet_contract_address::compute_deprecated_class_hash,
-        errors::state_errors::StateError,
-        transaction_hash::starknet_transaction_hash::calculate_declare_transaction_hash,
+        contract_address::compute_deprecated_class_hash, errors::state_errors::StateError,
+        transaction_hash::calculate_declare_transaction_hash,
     },
     definitions::{
         constants::VALIDATE_DECLARE_ENTRY_POINT_SELECTOR, general_config::StarknetGeneralConfig,
