+++ src/testing/mod.rs
@@ -1,3 +1,3 @@
-pub mod starknet_state;
-pub mod starknet_state_error;
+pub mod state;
+pub mod state_error;
 pub mod type_utils;
